ANK/requirements.txt
asgiref==3.8.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.17.0
gunicorn==23.0.0
packaging==24.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyotp==2.9.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.29.1
whitenoise==6.9.0
